adds unittests for DescriptionInputRenderer
